+++ packages/orq-rc/src/models/operations/createdatasource.ts
@@ -608,7 +608,7 @@ export const CreateDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KADNRAQ8933J9XAN2CN7YFST"),
+  _id: z.string().default("01KADVBBS7CZ4ED9QP03PEXA9H"),
   display_name: z.string(),
   description: z.string().optional(),
   status: CreateDatasourceStatus$inboundSchema,
@@ -651,7 +651,7 @@ export const CreateDatasourceResponseBody$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   CreateDatasourceResponseBody
 > = z.object({
-  id: z.string().default("01KADNRAQ8933J9XAN2CN7YFST"),
+  id: z.string().default("01KADVBBS7CZ4ED9QP03PEXA9H"),
   displayName: z.string(),
   description: z.string().optional(),
   status: CreateDatasourceStatus$outboundSchema,